Refactor AnswerPhoto to be a subcomponent of Answer
